+++ src/client-libs/index.js
@@ -2,21 +2,6 @@
 var reload_interval_ms = 10000;
 var backend_url = './backend.php';
 
-function generateRandomUsername() {
-    var username = "";
-    if (chance.bool()) {
-        username += chance.first();
-        if (chance.bool()) {
-            username += chance.last();
-        }
-    } else {
-        username += chance.word({syllables: 3})
-    }
-    if (chance.bool()) {
-        username += chance.integer({min: 30, max: 99});
-    }
-    return username.toLowerCase();
-}
 
 var app = angular.module('app', ["ngSanitize"]);
 
@@ -40,58 +25,82 @@ app.controller('MailboxController', ["$interval", "$http", "$log", function ($in
     var self = this;
 
     self.backend_url = backend_url;
+    self.public_config = {};
 
-    self.updateUsername = function (username) {
-        username = username.replace(/[@].*$/, ''); // remove part after "@"
-        if (self.username !== username) {
-            // changed
-            self.username = username;
-            hasher.setHash(self.username);
 
-            if (self.username.length > 0) {
-                self.address = self.username; // use username until real address has been loaded
-                self.updateMails();
-            } else {
-                self.randomize();
+    self.generateRandomAddress = function () {
+        var username = "";
+        if (chance.bool()) {
+            username += chance.first();
+            if (chance.bool()) {
+                username += chance.last();
             }
+        } else {
+            username += chance.word({syllables: 3})
         }
-        self.inputFieldUsername = self.username;
+        if (chance.bool()) {
+            username += chance.integer({min: 30, max: 99});
+        }
+        return username.toLowerCase() + "@" + chance.pick(self.public_config.domains, 1);
+    };
+
+    self.updateAddress = function (address) {
+        if (address.length === 0) {
+            self.randomize();
+        } else {
+            if (self.address !== address) {
+                // changed
+                if (address.indexOf("@") === -1) {
+                    address += "@" + chance.pick(self.public_config.domains, 1);
+                }
+                self.address = address;
+                hasher.setHash(self.address);
+                self.updateMails();
+            }
+            self.inputFieldAddress = self.address;
+        }
     };
 
 
     self.randomize = function () {
-        self.updateUsername(generateRandomUsername());
+        self.updateAddress(self.generateRandomAddress());
     };
 
 
     self.onHashChange = function (hash) {
-        self.updateUsername(hash);
+        self.updateAddress(hash);
     };
 
     self.$onInit = function () {
+        self.loadConfig().then(function () {
+            self.afterLoadConfig();
+        });
+    };
+
+    self.afterLoadConfig = function () {
         hasher.changed.add(self.onHashChange.bind(self));
         hasher.initialized.add(self.onHashChange.bind(self)); //add initialized listener (to grab initial value in case it is already set)
         hasher.init(); //initialize hasher (start listening for history changes)
 
         $interval(self.updateMails, reload_interval_ms);
+        self.updateMails()
     };
 
     self.updateMails = function () {
-        if (self.username) {
-            self.loadEmailsAsync(self.username);
+        if (self.address) {
+            self.loadEmailsAsync(self.address);
         }
     };
 
-    self.loadEmailsAsync = function (username) {
-        $http.get(backend_url, {params: {username: username}})
+    self.loadEmailsAsync = function (address) {
+        $http.get(backend_url, {params: {address: address}})
             .then(function successCallback(response) {
                 if (response.data.mails) {
                     self.error = null;
                     self.mails = response.data.mails;
                     self.address = response.data.address;
-                    self.username = response.data.username;
-                    if (self.inputFieldUsername === self.username) {
-                        self.inputFieldUsername = self.address;
+                    if (self.inputFieldAddress === self.address) {
+                        self.inputFieldAddress = self.address;
                     }
                 } else {
                     self.error = {
@@ -111,13 +120,27 @@ app.controller('MailboxController', ["$interval", "$http", "$log", function ($in
             });
     };
 
+    self.loadConfig = function () {
+        return $http.get(backend_url, {params: {get_config: true}})
+            .then(function successCallback(response) {
+                self.public_config = response.data.config;
+            }, function errorCallback(response) {
+                $log.error(response, this);
+                self.error = {
+                    title: "HTTP_ERROR",
+                    desc: "There is a problem with loading the config. (HTTP_ERROR).",
+                    detail: response
+                };
+            });
+    };
+
     self.deleteMail = function (mail, index) {
         // instantly remove from frontend.
         self.mails.splice(index, 1);
 
         // remove on backend.
         var firstTo = Object.keys(mail.to)[0];
-        $http.get(backend_url, {params: {username: firstTo, delete_email_id: mail.id}})
+        $http.get(backend_url, {params: {address: firstTo, delete_email_id: mail.id}})
             .then(
                 function successCallback(response) {
                     self.updateMails();
@@ -131,7 +154,4 @@ app.controller('MailboxController', ["$interval", "$http", "$log", function ($in
                     };
                 });
     };
-
-    // Initial load
-    self.updateMails()
 }]);